[Clang importer] Wire up extensions for globals-as-members.
Wire up the extensions created when importing globals-as-members via
the same mechanisms we use for Objective-C categories, e.g.,
implementing loadAllExtensions() to find the extensions and lazily
loading their members via the member loader. Addresses most of my
comments on the way extensions were wired in.

Extend our testing for importing globals as members to two different
submodules, so we can see the separation of extensions.

As part of this, fold getOrCreateExtension into importDeclContextOf.

+++ test/IDE/Inputs/custom-modules/ImportAsMember.h
@@ -5,9 +5,10 @@ struct __attribute__((swift_name("Struct1"))) IAMStruct1 {
   double x, y, z;
 };
 
-extern int IAMStruct1GlobalVar __attribute__((swift_name("Struct1.globalVar")));
+extern double IAMStruct1GlobalVar
+    __attribute__((swift_name("Struct1.globalVar")));
 
-extern struct IAMStruct1 IAMStruct1CreateSimple(int value)
+extern struct IAMStruct1 IAMStruct1CreateSimple(double value)
     __attribute__((swift_name("Struct1.init(value:)")));
 
 extern double IAMStruct1GetRadius(struct IAMStruct1 s)
